move more similar methods to the base impls
