+++ src/c/transport/udp.h
@@ -11,12 +11,8 @@ extern "C" { // only need to export C interface if
 #include "transport.h"
 #include <netinet/in.h>
 #include <stdint.h>
-#ifdef ENABLE_TLS
-#include <wolfssl/options.h>
-#include <wolfssl/ssl.h>
-#endif
 
-#define IPV4_STR_LEN 16
+void handle_tls_mode(rasta_transport_socket *transport_socket);
 
 /**
  * This function will initialise an udp socket and return its file descriptor, which is used to reference it in later
@@ -67,10 +63,12 @@ void udp_send(rasta_transport_channel *transport_channel, unsigned char *message
 void udp_send_sockaddr(rasta_transport_channel *transport_channel, unsigned char *message, size_t message_len, struct sockaddr_in receiver);
 
 /**
- * Closes the udp socket
- * @param transport_socket the rasta_transport_socket which identifies the socket
+ * Closes the udp channel
+ * @param transport_channel the rasta_transport_channel which identifies the channel
  */
-void udp_close(rasta_transport_socket *transport_socket);
+// TODO: This is unused! Why? Why does udp_close even exist?
+// TODO: transport_channel signature needed to compile, but doesn't make sense here!
+void udp_close(rasta_transport_channel *transport_channel);
 
 #ifdef __cplusplus
 }
